fix: Before didnt work, this should work both ssr and non-ssr
